[udev][usb-xhci] Implement scatter/gather support
The usb-xhci driver is now able to handle non-contiguous data buffers
for USB transactions.

As part of this change, we also clean up how iotxns are queued.
Removed the clunky xhci_process_deferred_txns() mechanism, replacing it with
a pipeline with separate lists for transactions waiting to begin processing
and transactions waiting to be completed.

Change-Id: I614ddc99c2abb0424143677313a8eb2885783b77